+++ service/controller/resource/app/current.go
@@ -22,9 +22,11 @@ func (r *Resource) GetCurrentState(ctx context.Context, obj interface{}) ([]*v1a
 		return nil, microerror.Mask(err)
 	}
 
-	// Apps are deleted by the provider operator when it deletes
-	// the tenant cluster namespace in the control plane cluster.
-	if key.IsDeleted(objectMeta) {
+	// Apps are deleted by the provider operator when it deletes the tenant
+	// cluster namespace in the control plane cluster. Except on KVM where
+	// the app-operator app needs to be deleted so all pods in the cluster
+	// namespace are deleted.
+	if key.IsDeleted(objectMeta) && r.provider != "kvm" {
 		r.logger.LogCtx(ctx, "level", "debug", "message", "redirecting app deletion to provider operators")
 		r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
 		resourcecanceledcontext.SetCanceled(ctx)