+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.25.1] - 2021-03-17
+## [0.26.0] - 2021-03-25
 
 ### Added
 
 - Assign app catalog name from the component in release CR.
+
+## [0.25.1] - 2021-03-17
 
 ### Fixed
 
@@ -99,7 +101,8 @@ Replacement for [0.23.21] because of incorrect `bundleVersion` in [0.23.21]
 
 - Added GitHub workflows.
 
-[Unreleased]: https://github.com/giantswarm/cluster-operator/compare/v0.25.1...HEAD
+[Unreleased]: https://github.com/giantswarm/cluster-operator/compare/v0.26.0...HEAD
+[0.26.0]: https://github.com/giantswarm/cluster-operator/compare/v0.25.1...v0.26.0
 [0.25.1]: https://github.com/giantswarm/cluster-operator/compare/v0.25.0...v0.25.1
 [0.25.0]: https://github.com/giantswarm/cluster-operator/compare/v0.24.2...v0.25.0
 [0.24.2]: https://github.com/giantswarm/cluster-operator/compare/v0.24.1...v0.24.2